Move findline class
